requirements.txt
flask>=2.3.0
requests>=2.31.0
mysql-connector-python>=8.0.32
DrissionPage>=4.0.0
